+++ _delb/xpath/__init__.py
@@ -65,6 +65,7 @@ from __future__ import annotations
 
 from functools import lru_cache
 from typing import TYPE_CHECKING, Optional
+from warnings import warn
 
 # DROPWITH Python 3.8 and replace w/ imports from collections.abc
 from typing import Collection, Iterable, Iterator, Mapping, Sequence
@@ -170,6 +171,10 @@ def evaluate(
 ) -> QueryResults:
     # global namespaces are guaranteed by the Namespaces implementation
     if namespaces is None:
+        warn(
+            "Default namespace declarations that are carried over won't be available "
+            "in future versions. The declarations need to be passed explicitly."
+        )
         _namespaces = node.namespaces
     elif isinstance(namespaces, Namespaces):
         # b/c it would break fallback chains