+++ tests/test_attributes.py
@@ -130,6 +130,8 @@ def test_attribute_object():
     assert attribute.value == "detached"
 
 
+# TODO remove warning filter when empty declarations are used as fallback
+@pytest.mark.filterwarnings("ignore:.* namespace declarations")
 def test_comparison():
     document = Document(
         dedent(
@@ -178,9 +180,9 @@ def test_comparison():
 
 def test_delete_namespaced_attribute():
     root = Document('<root><node xmlns:p="ns" p:a="1" p:b="2"/></root>').root
-    node = root.css_select("node").first
+    node = root.css_select("node", namespaces={None: ""}).first
     assert len(node.attributes) == 2
-    del node.attributes["ns":"a"]
+    del node.attributes[("ns", "a")]
     assert len(node.attributes) == 1