+++ tests/test_tag_node.py
@@ -1,4 +1,5 @@
 from copy import copy, deepcopy
+from typing import Final
 
 import pytest
 
@@ -16,6 +17,8 @@ from delb.exceptions import InvalidOperation
 from tests.conftest import XML_FILES
 
 from tests.utils import assert_nodes_are_in_document_order, skip_long_running_test
+
+TEI_NAMESPACE: Final = "http://www.tei-c.org/ns/1.0"
 
 
 def is_pagebreak(node):
@@ -163,6 +166,8 @@ def test_detach_and_document_property():
     assert str(document) == '<?xml version="1.0" encoding="UTF-8"?><root/>'
 
 
+# TODO remove warning filter when empty declarations are used as fallback
+@pytest.mark.filterwarnings("ignore:.* namespace declarations")
 def test_detach_node_with_tail_1():
     root = Document("<root><a>c<c/></a>b<d/></root>").root
 
@@ -216,6 +221,8 @@ def test_detach_node_retain_child_nodes():
         root.detach(retain_child_nodes=True)
 
 
+# TODO remove warning filter when empty declarations are used as fallback
+@pytest.mark.filterwarnings("ignore:.* namespace declarations")
 def test_detach_node_retains_namespace_prefixes():
     # libxml2 loses the notion if a default prefix for nodes that have been
     # removed from a parent node
@@ -306,7 +313,9 @@ def test_getitem():
 
 def test_id_property(files_path):
     document = Document(files_path / "marx_manifestws_1848.TEI-P5.xml")
-    publisher = document.css_select("publicationStmt publisher").first
+    publisher = document.css_select(
+        "publicationStmt publisher", namespaces={None: TEI_NAMESPACE}
+    ).first
 
     assert publisher.id == "DTACorpusPublisher"
 
@@ -437,6 +446,8 @@ def test_last_descendant():
     assert c.last_descendant is None
 
 
+# TODO remove warning filter when empty declarations are used as fallback
+@pytest.mark.filterwarnings("ignore:.* namespace declarations")
 @skip_long_running_test
 @pytest.mark.parametrize("file", XML_FILES)
 def test_location_path_and_xpath_concordance(file):